+++ src/main/java/com/woowahan/racing/model/GameRandom.java
@@ -1,4 +1,4 @@
-package com.woowahan.racing.util;
+package com.woowahan.racing.model;
 
 import java.util.Random;
